+++ js/modules/get-random-array-element.js
@@ -1,0 +1,5 @@
+import {returnRundomNumber} from './modules/return-rundom-number.js';
+
+const getRandomArrayElement = (elements) => elements[returnRundomNumber(0, elements.length - 1)];
+
+export {getRandomArrayElement};
